#16 Verify the existence of third party files before packaging.

+++ kde/install.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
-kpackagetool5 -t Plasma/Applet --install plasmoid
-kpackagetool5 -t Plasma/Applet --upgrade plasmoid
 
+if [ -f "../third_party/hsluv-glsl/hsluv-glsl.fsh" ];then
+    kpackagetool5 -t Plasma/Applet --install plasmoid
+    kpackagetool5 -t Plasma/Applet --upgrade plasmoid
+else
+    echo "Cannot find third party files."
+fi
